+++ src/includes/fdc.h
@@ -33,7 +33,7 @@ extern void	FDC_DMA_FIFO_Push ( Uint8 Byte );
 extern Uint8	FDC_DMA_FIFO_Pull ( void );
 
 extern void	FDC_Buffer_Reset ( void );
-extern void	FDC_Buffer_Add_Timing ( Uint8 Byte , int Timing );
+extern void	FDC_Buffer_Add_Timing ( Uint8 Byte , Uint16 Timing );
 extern void	FDC_Buffer_Add ( Uint8 Byte );
 extern Uint16	FDC_Buffer_Read_Timing ( void );
 extern Uint8	FDC_Buffer_Read_Byte ( void );
